Changed name of commands folder to command and moved DeleteCommands.ts to command folder
