+++ src/command/RegisterDiscussionCourse.ts
